Local analysis
