+++ datavec/datavec-api/src/main/java/org/datavec/api/transform/analysis/counter/DoubleAnalysisCounter.java
@@ -14,23 +14,23 @@
  *  *    limitations under the License.
  */
 
-package org.datavec.spark.transform.analysis.columns;
+package org.datavec.api.transform.analysis.counter;
 
 import com.tdunning.math.stats.TDigest;
 import lombok.AllArgsConstructor;
 import lombok.Data;
-import org.apache.spark.util.StatCounter;
+import org.datavec.api.transform.analysis.AnalysisCounter;
 import org.datavec.api.writable.Writable;
-import org.datavec.spark.transform.analysis.AnalysisCounter;
+import org.datavec.local.transforms.analysis.statcounter.StatCounter;
 
 /**
- * A counter function for doing analysis on Long columns, on Spark
+ * A counter function for doing analysis on Double columns, on Spark
  *
  * @author Alex Black
  */
 @AllArgsConstructor
 @Data
-public class LongAnalysisCounter implements AnalysisCounter<LongAnalysisCounter> {
+public class DoubleAnalysisCounter implements AnalysisCounter<DoubleAnalysisCounter> {
 
     private StatCounter counter = new StatCounter();
     private long countZero = 0;
@@ -38,52 +38,57 @@ public class LongAnalysisCounter implements AnalysisCounter<LongAnalysisCounter>
     private long countMaxValue = 0;
     private long countPositive = 0;
     private long countNegative = 0;
-    /**
-     * A histogram structure that will record a sketch of a distribution.
-     *
-     * The compression argument regulates how accuracy should be traded for size? A value of N here
-     * will give quantile errors almost always less than 3/N with considerably smaller errors expected
-     * for extreme quantiles. Conversely, you should expect to track about 5 N centroids for this
-     * accuracy.
-     */
-    private TDigest digest = TDigest.createDigest(100);
+    private long countNaN = 0;
+  /**
+   * A histogram structure that will record a sketch of a distribution.
+   *
+   * The compression argument regulates how accuracy should be traded for size? A value of N here
+   * will give quantile errors almost always less than 3/N with considerably smaller errors expected
+   * for extreme quantiles. Conversely, you should expect to track about 5 N centroids for this
+   * accuracy.
+   */
+  private TDigest digest = TDigest.createDigest(100);
 
-    public LongAnalysisCounter() {};
 
-    public long getMinValueSeen() {
-        return (long) counter.min();
+    public DoubleAnalysisCounter() {};
+
+    public double getMinValueSeen() {
+        return counter.getMin();
     };
 
-    public long getMaxValueSeen() {
-        return (long) counter.max();
+    public double getMaxValueSeen() {
+        return counter.getMax();
     };
 
-    public long getSum() {
-        return (long) counter.sum();
+    public double getSum() {
+        return counter.getSum();
     };
 
     public long getCountTotal() {
-        return counter.count();
+        return counter.getCount();
     };
 
     public double getSampleStdev() {
-        return counter.sampleStdev();
+        return counter.getStddev(false);
     };
 
     public double getMean() {
-        return counter.mean();
+        return counter.getMean();
     }
 
     public double getSampleVariance() {
-        return counter.sampleVariance();
+        return counter.getVariance(false);
     }
 
     @Override
-    public LongAnalysisCounter add(Writable writable) {
-        long value = writable.toLong();
+    public DoubleAnalysisCounter add(Writable writable) {
+        double value = writable.toDouble();
 
         if (value == 0)
             countZero++;
+
+        if (value == Double.NaN)
+            countNaN++;
 
         if (value == getMinValueSeen())
             countMinValue++;
@@ -103,26 +108,26 @@ public class LongAnalysisCounter implements AnalysisCounter<LongAnalysisCounter>
             countNegative++;
         } ;
 
-        digest.add((double) value);
-        counter.merge((double) value);
+        digest.add(value);
+        counter.add(value);
 
         return this;
     }
 
-    public LongAnalysisCounter merge(LongAnalysisCounter other) {
-        long otherMin = other.getMinValueSeen();
+    public DoubleAnalysisCounter merge(DoubleAnalysisCounter other) {
+        double otherMin = other.getMinValueSeen();
         long newCountMinValue;
         if (getMinValueSeen() == otherMin) {
             newCountMinValue = countMinValue + other.getCountMinValue();
         } else if (getMinValueSeen() > otherMin) {
-            //Keep other, take count from other
+            //Keep other, take count from othergetSampleStdDev
             newCountMinValue = other.getCountMinValue();
         } else {
             //Keep this min, no change to count
             newCountMinValue = countMinValue;
         }
 
-        long otherMax = other.getMaxValueSeen();
+        double otherMax = other.getMaxValueSeen();
         long newCountMaxValue;
         if (getMaxValueSeen() == otherMax) {
             newCountMaxValue = countMaxValue + other.getCountMaxValue();
@@ -136,8 +141,8 @@ public class LongAnalysisCounter implements AnalysisCounter<LongAnalysisCounter>
 
         digest.add(other.getDigest());
 
-        return new LongAnalysisCounter(counter.merge(other.getCounter()), countZero + other.getCountZero(),
+        return new DoubleAnalysisCounter(counter.merge(other.getCounter()), countZero + other.getCountZero(),
                         newCountMinValue, newCountMaxValue, countPositive + other.getCountPositive(),
-                        countNegative + other.getCountNegative(), digest);
+                        countNegative + other.getCountNegative(), countNaN + other.getCountNaN(), digest);
     }
 }
