+++ datavec/datavec-api/src/main/java/org/datavec/api/transform/analysis/counter/NDArrayAnalysisCounter.java
@@ -14,12 +14,12 @@
  *  *    limitations under the License.
  */
 
-package org.datavec.spark.transform.analysis.columns;
+package org.datavec.api.transform.analysis.counter;
 
+import org.datavec.api.transform.analysis.AnalysisCounter;
 import org.datavec.api.transform.analysis.columns.NDArrayAnalysis;
 import org.datavec.api.writable.NDArrayWritable;
 import org.datavec.api.writable.Writable;
-import org.datavec.spark.transform.analysis.AnalysisCounter;
 import org.nd4j.linalg.api.ndarray.INDArray;
 
 import java.util.HashMap;
