+++ datavec/datavec-api/src/main/java/org/datavec/api/transform/analysis/counter/StringAnalysisMergeFunction.java
@@ -1,0 +1,29 @@
+/*-
+ *  * Copyright 2016 Skymind, Inc.
+ *  *
+ *  *    Licensed under the Apache License, Version 2.0 (the "License");
+ *  *    you may not use this file except in compliance with the License.
+ *  *    You may obtain a copy of the License at
+ *  *
+ *  *        http://www.apache.org/licenses/LICENSE-2.0
+ *  *
+ *  *    Unless required by applicable law or agreed to in writing, software
+ *  *    distributed under the License is distributed on an "AS IS" BASIS,
+ *  *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  *    See the License for the specific language governing permissions and
+ *  *    limitations under the License.
+ */
+
+package org.datavec.api.transform.analysis.counter;
+
+import java.util.function.BiFunction;
+
+/**
+ * Created by Alex on 5/03/2016.
+ */
+public class StringAnalysisMergeFunction
+                implements BiFunction<StringAnalysisCounter, StringAnalysisCounter, StringAnalysisCounter> {
+    public StringAnalysisCounter apply(StringAnalysisCounter v1, StringAnalysisCounter v2) {
+        return v1.merge(v2);
+    }
+}
