+++ datavec/datavec-api/src/main/java/org/datavec/api/transform/analysis/histogram/DoubleHistogramCounter.java
@@ -14,35 +14,35 @@
  *  *    limitations under the License.
  */
 
-package org.datavec.spark.transform.analysis.histogram;
+package org.datavec.api.transform.analysis.histogram;
 
 import org.datavec.api.writable.Writable;
 
 /**
- * A counter for building histograms (of String length) on a String column
+ * A counter for building histograms on a Double column
  *
  * @author Alex Black
  */
-public class StringHistogramCounter implements HistogramCounter {
+public class DoubleHistogramCounter implements HistogramCounter {
 
-    private final int minLength;
-    private final int maxLength;
+    private final double minValue;
+    private final double maxValue;
     private final int nBins;
     private final double[] bins;
     private final long[] binCounts;
 
-    public StringHistogramCounter(int minLength, int maxLength, int nBins) {
-        this.minLength = minLength;
-        this.maxLength = maxLength;
+    public DoubleHistogramCounter(double minValue, double maxValue, int nBins) {
+        this.minValue = minValue;
+        this.maxValue = maxValue;
         this.nBins = nBins;
 
         bins = new double[nBins + 1]; //+1 because bins are defined by a range of values: bins[i] to bins[i+1]
-        double step = ((double) (maxLength - minLength)) / nBins;
+        double step = (maxValue - minValue) / nBins;
         for (int i = 0; i < bins.length; i++) {
             if (i == bins.length - 1)
-                bins[i] = maxLength;
+                bins[i] = maxValue;
             else
-                bins[i] = i * step;
+                bins[i] = minValue + i * step;
         }
 
         binCounts = new long[nBins];
@@ -51,12 +51,12 @@ public class StringHistogramCounter implements HistogramCounter {
 
     @Override
     public HistogramCounter add(Writable w) {
-        double d = w.toString().length();
+        double d = w.toDouble();
 
         //Not super efficient, but linear search on 20-50 items should be good enough
         int idx = -1;
         for (int i = 0; i < nBins; i++) {
-            if (d >= bins[i] && d < bins[i]) {
+            if (d >= bins[i] && d < bins[i + 1]) {
                 idx = i;
                 break;
             }
@@ -70,17 +70,17 @@ public class StringHistogramCounter implements HistogramCounter {
     }
 
     @Override
-    public StringHistogramCounter merge(HistogramCounter other) {
+    public DoubleHistogramCounter merge(HistogramCounter other) {
         if (other == null)
             return this;
-        if (!(other instanceof StringHistogramCounter))
+        if (!(other instanceof DoubleHistogramCounter))
             throw new IllegalArgumentException("Cannot merge " + other);
 
-        StringHistogramCounter o = (StringHistogramCounter) other;
+        DoubleHistogramCounter o = (DoubleHistogramCounter) other;
 
-        if (minLength != o.minLength || maxLength != o.maxLength)
-            throw new IllegalStateException("Min/max values differ: (" + minLength + "," + maxLength + ") " + " vs. ("
-                            + o.minLength + "," + o.maxLength + ")");
+        if (minValue != o.minValue || maxValue != o.maxValue)
+            throw new IllegalStateException("Min/max values differ: (" + minValue + "," + maxValue + ") " + " vs. ("
+                            + o.minValue + "," + o.maxValue + ")");
         if (nBins != o.nBins)
             throw new IllegalStateException("Different number of bins: " + nBins + " vs " + o.nBins);
