+++ datavec/datavec-api/src/main/java/org/datavec/api/transform/analysis/histogram/NDArrayHistogramCounter.java
@@ -14,7 +14,7 @@
  *  *    limitations under the License.
  */
 
-package org.datavec.spark.transform.analysis.histogram;
+package org.datavec.api.transform.analysis.histogram;
 
 import org.datavec.api.writable.DoubleWritable;
 import org.datavec.api.writable.NDArrayWritable;
