+++ datavec/datavec-spark/src/main/java/org/datavec/spark/transform/analysis/histogram/HistogramAddFunction.java
@@ -19,6 +19,7 @@ package org.datavec.spark.transform.analysis.histogram;
 import lombok.AllArgsConstructor;
 import org.apache.spark.api.java.function.Function2;
 import org.datavec.api.transform.ColumnType;
+import org.datavec.api.transform.analysis.histogram.*;
 import org.datavec.api.transform.metadata.CategoricalMetaData;
 import org.datavec.api.transform.schema.Schema;
 import org.datavec.api.writable.Writable;
